+++ Nukepayload2.Csv.TestCommon/RawModel.vb
@@ -7,3 +7,12 @@
     Public Property SingleValue As Single
     Public Property LongValue As Long
 End Class
+
+Public Class NullableModel
+    Public Property IntegerValue As Integer?
+    Public Property DoubleValue As Double?
+    Public Property DateValue As Date?
+    Public Property BooleanValue As Boolean?
+    Public Property SingleValue As Single?
+    Public Property LongValue As Long?
+End Class