+++ Nukepayload2.Csv/Formatters/CsvBooleanFormatter.vb
@@ -2,23 +2,22 @@
     Inherits Singleton(Of CsvBooleanFormatter)
     Implements ICsvRecordFormatter
 
+    Private Const AscIIToUpperMask As Integer = &HFFDF
+
     Public Function Parse(text As StringSegment) As Object Implements ICsvRecordFormatter.Parse
-        If text.IsNullOrEmpty Then
-            Throw New FormatException("Boolean expected.")
-        End If
         If text.Length = 4 Then
-            If (AscW(text(0)) And &HFFDF) = AscW("T"c) AndAlso
-               (AscW(text(1)) And &HFFDF) = AscW("R"c) AndAlso
-               (AscW(text(2)) And &HFFDF) = AscW("U"c) AndAlso
-               (AscW(text(3)) And &HFFDF) = AscW("E"c) Then
+            If (Convert.ToInt32(text(0)) And AscIIToUpperMask) = Convert.ToInt32("T"c) AndAlso
+               (Convert.ToInt32(text(1)) And AscIIToUpperMask) = Convert.ToInt32("R"c) AndAlso
+               (Convert.ToInt32(text(2)) And AscIIToUpperMask) = Convert.ToInt32("U"c) AndAlso
+               (Convert.ToInt32(text(3)) And AscIIToUpperMask) = Convert.ToInt32("E"c) Then
                 Return True
             End If
         ElseIf text.Length = 5 Then
-            If (AscW(text(0)) And &HFFDF) = AscW("F"c) AndAlso
-               (AscW(text(1)) And &HFFDF) = AscW("A"c) AndAlso
-               (AscW(text(2)) And &HFFDF) = AscW("L"c) AndAlso
-               (AscW(text(3)) And &HFFDF) = AscW("S"c) AndAlso
-               (AscW(text(4)) And &HFFDF) = AscW("E"c) Then
+            If (Convert.ToInt32(text(0)) And AscIIToUpperMask) = Convert.ToInt32("F"c) AndAlso
+               (Convert.ToInt32(text(1)) And AscIIToUpperMask) = Convert.ToInt32("A"c) AndAlso
+               (Convert.ToInt32(text(2)) And AscIIToUpperMask) = Convert.ToInt32("L"c) AndAlso
+               (Convert.ToInt32(text(3)) And AscIIToUpperMask) = Convert.ToInt32("S"c) AndAlso
+               (Convert.ToInt32(text(4)) And AscIIToUpperMask) = Convert.ToInt32("E"c) Then
                 Return False
             End If
         End If
@@ -28,8 +27,41 @@
     Public Function GetString(data As Object, format As String) As String Implements ICsvRecordFormatter.GetString
         Return DirectCast(data, Boolean).ToString
     End Function
+End Class
 
-    Public Function ParseBlittablePrimitive(text As StringSegment, ByRef primitive As CsvBlittablePrimitive) As Boolean Implements ICsvRecordFormatter.ParseBlittablePrimitive
-        Return False
+Friend Class CsvBooleanNullableFormatter
+    Inherits Singleton(Of CsvBooleanNullableFormatter)
+    Implements ICsvRecordFormatter
+
+    Private Const AscIIToUpperMask As Integer = &HFFDF
+
+    Public Function Parse(text As StringSegment) As Object Implements ICsvRecordFormatter.Parse
+        If text.IsNullOrEmpty Then
+            Return Nothing
+        End If
+        If text.Length = 4 Then
+            If (Convert.ToInt32(text(0)) And AscIIToUpperMask) = Convert.ToInt32("T"c) AndAlso
+               (Convert.ToInt32(text(1)) And AscIIToUpperMask) = Convert.ToInt32("R"c) AndAlso
+               (Convert.ToInt32(text(2)) And AscIIToUpperMask) = Convert.ToInt32("U"c) AndAlso
+               (Convert.ToInt32(text(3)) And AscIIToUpperMask) = Convert.ToInt32("E"c) Then
+                Return True
+            End If
+        ElseIf text.Length = 5 Then
+            If (Convert.ToInt32(text(0)) And AscIIToUpperMask) = Convert.ToInt32("F"c) AndAlso
+               (Convert.ToInt32(text(1)) And AscIIToUpperMask) = Convert.ToInt32("A"c) AndAlso
+               (Convert.ToInt32(text(2)) And AscIIToUpperMask) = Convert.ToInt32("L"c) AndAlso
+               (Convert.ToInt32(text(3)) And AscIIToUpperMask) = Convert.ToInt32("S"c) AndAlso
+               (Convert.ToInt32(text(4)) And AscIIToUpperMask) = Convert.ToInt32("E"c) Then
+                Return False
+            End If
+        End If
+        Throw New FormatException("Boolean expected.")
+    End Function
+
+    Public Function GetString(data As Object, format As String) As String Implements ICsvRecordFormatter.GetString
+        If data Is Nothing Then
+            Return Nothing
+        End If
+        Return DirectCast(data, Boolean).ToString
     End Function
 End Class