+++ Nukepayload2.Csv/Formatters/CsvDoubleNullableFormatter.vb
@@ -1,12 +1,15 @@
 ﻿Imports Nukepayload2.Csv.Numbers.Double
 
-Friend Class CsvDoubleFormatter
-    Inherits Singleton(Of CsvDoubleFormatter)
+Friend Class CsvDoubleNullableFormatter
+    Inherits Singleton(Of CsvDoubleNullableFormatter)
     Implements ICsvRecordFormatter
 
     Private ReadOnly _doubleView As New DoubleSemanticView
 
     Public Function Parse(text As StringSegment) As Object Implements ICsvRecordFormatter.Parse
+        If text.IsNullOrEmpty Then
+            Return Nothing
+        End If
         Dim value As Double = Nothing
         If _doubleView.TryParse(text, value) Then
             Return value
@@ -15,7 +18,9 @@ Friend Class CsvDoubleFormatter
     End Function
 
     Public Function GetString(data As Object, format As String) As String Implements ICsvRecordFormatter.GetString
+        If data Is Nothing Then
+            Return Nothing
+        End If
         Return DirectCast(data, Double).ToString(format)
     End Function
-
 End Class