+++ Nukepayload2.Csv/Formatters/CsvIntegerNullableFormatter.vb
@@ -1,12 +1,18 @@
-﻿Friend Class CsvIntegerFormatter
-    Inherits Singleton(Of CsvIntegerFormatter)
+﻿Friend Class CsvIntegerNullableFormatter
+    Inherits Singleton(Of CsvIntegerNullableFormatter)
     Implements ICsvRecordFormatter
 
     Public Function Parse(text As StringSegment) As Object Implements ICsvRecordFormatter.Parse
+        If text.IsNullOrEmpty Then
+            Return Nothing
+        End If
         Return Integer.Parse(text.CopyToString)
     End Function
 
     Public Function GetString(data As Object, format As String) As String Implements ICsvRecordFormatter.GetString
+        If data Is Nothing Then
+            Return Nothing
+        End If
         Return DirectCast(data, Integer).ToString(format)
     End Function
 End Class