+++ Nukepayload2.Csv/Formatters/CsvLongNullableFormatter.vb
@@ -1,12 +1,18 @@
-﻿Friend Class CsvLongFormatter
-    Inherits Singleton(Of CsvLongFormatter)
+﻿Friend Class CsvLongNullableFormatter
+    Inherits Singleton(Of CsvLongNullableFormatter)
     Implements ICsvRecordFormatter
 
     Public Function Parse(text As StringSegment) As Object Implements ICsvRecordFormatter.Parse
+        If text.IsNullOrEmpty Then
+            Return Nothing
+        End If
         Return Long.Parse(text.CopyToString)
     End Function
 
     Public Function GetString(data As Object, format As String) As String Implements ICsvRecordFormatter.GetString
+        If data Is Nothing Then
+            Return Nothing
+        End If
         Return DirectCast(data, Long).ToString(format)
     End Function
 End Class