+++ Nukepayload2.Csv/Formatters/CsvTextFormatter.vb
@@ -9,8 +9,4 @@
     Public Function GetString(data As Object, format As String) As String Implements ICsvRecordFormatter.GetString
         Return DirectCast(data, String)
     End Function
-
-    Public Function ParseBlittablePrimitive(text As StringSegment, ByRef primitive As CsvBlittablePrimitive) As Boolean Implements ICsvRecordFormatter.ParseBlittablePrimitive
-        Return False
-    End Function
 End Class